Reverted multi tick cache - not needed!
My misunderstanding of the reconcile process - it was only using the last filled item in the cache during reconcile anyway. So the reverted predictedobjectcache is now mostly like the fishnet orig, only difference being it has the double update fix (fishnet pending that fix)  and it also does not set the rigidbody state when it receives the new state - but always waits to do it until reconcile - which I think helps a bit when the car hits the spheres.